Hibernate

+++ src/main/java/dao/GoodDao.java
@@ -43,7 +43,7 @@ public class GoodDao {
             ResultSet resultSet = preparedStatement.executeQuery();
             logger.debug(sql);
             if (resultSet.next()) {
-                long idDb = resultSet.getLong(1);
+                int idDb = resultSet.getInt(1);
                 String name = resultSet.getString(2);
                 String description = resultSet.getString(3);
                 Double price = resultSet.getDouble(4);
